iter3: improved http router
