+++ go.mod
@@ -3,6 +3,7 @@ module github.com/ElfAstAhe/url-shortener
 go 1.24.5
 
 require (
+	github.com/go-chi/chi/v5 v5.2.2
 	github.com/google/uuid v1.6.0
 	github.com/stretchr/testify v1.10.0
 )
