+++ internal/config/config.go
@@ -8,10 +8,10 @@ const DBKindInMemory string = "IN_MEMORY"
 const DBKindPostgres string = "POSTGRES"
 
 type Config struct {
-	AppName string       `json:"app_name"`
-	BaseURL string       `json:"base_url"`
-	HTTP    []HTTPConfig `json:"http"`
-	DB      DBConfig     `json:"db"`
+	AppName string     `json:"app_name"`
+	BaseURL string     `json:"base_url"`
+	HTTP    HTTPConfig `json:"http"`
+	DB      DBConfig   `json:"db"`
 }
 
 var GlobalConfig Config
@@ -21,7 +21,7 @@ func (c *Config) LoadConfig() error {
 
 	c.AppName = "URL shortener"
 	c.BaseURL = "http://localhost:8080"
-	c.HTTP = []HTTPConfig{*NewHTTPConfig(DefaultHTTPSchema, DefaultHTTPHost, DefaultHTTPPort)}
+	c.HTTP = *NewHTTPConfig(DefaultHTTPSchema, DefaultHTTPHost, DefaultHTTPPort)
 	c.DB = *NewDBConfig(DBKindInMemory)
 
 	return nil
